deployment/dump_function/requirements.txt
requests==2.28.2
google-cloud-storage==2.7.0
